Move `Timer` to `timeutils`.
